moves non page file out of page directory
