+++ content/checklists/remoteReady.ts
